feat(suite): extend tooltip content of account item when sidebar collapsed
- Always display account name (e.g. Ethereum #1)
- Display account item type (staking, tokens)
- Display non-default account type (e.g. Ledger)
- Plus do some basic component refactor to make it a little bit more readable (place styles to seperate files, do some component de-composition, etc.)
