+++ packages/suite/src/components/wallet/WalletLayout/AccountsMenu/AccountItem/AccountRow/AccountItemContent/index.ts
@@ -1,0 +1,1 @@
+export * from './AccountItemContent';
