+++ scripts/circular-dependencies/madge-snapshot.txt
@@ -155,7 +155,6 @@ packages/suite/src/components/wallet/Fees/StandardFee/StandardFee.tsx > packages
 packages/suite/src/components/wallet/Fees/StandardFee/StandardFee.tsx > packages/suite/src/components/wallet/Fees/StandardFee/EthereumFeeCards.tsx
 packages/suite/src/components/wallet/Fees/StandardFee/StandardFee.tsx > packages/suite/src/components/wallet/Fees/StandardFee/MiscFeeCards.tsx
 packages/suite/src/components/wallet/TransactionItem/TransactionTarget/TransactionTargetsList.tsx > packages/suite/src/components/wallet/TransactionItem/TransactionTarget/TransactionTarget.tsx
-packages/suite/src/components/wallet/WalletLayout/AccountsMenu/AccountItem/AccountItem.tsx > packages/suite/src/components/wallet/WalletLayout/AccountsMenu/AccountItem/AccountRow.tsx
 packages/suite/src/utils/wallet/graph/utils.ts > packages/suite/src/utils/wallet/graph/utilsShared.ts > packages/suite/src/utils/wallet/graph/utilsWorker.ts
 packages/suite/src/utils/wallet/graph/utilsShared.ts > packages/suite/src/utils/wallet/graph/utilsWorker.ts
 packages/suite/src/views/dashboard/AssetsView/AssetTable/AssetTable.tsx > packages/suite/src/views/dashboard/AssetsView/AssetsView.tsx
